% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BMTplat}
\alias{BMTplat}
\title{Bone Marrow Transplant Registry}
\format{
A data frame with 408 rows and 3 variables:
\describe{
\item{time}{time to event since transplant (in months)}
\item{status}{event status, 1 is dead from treatment related causes, 2 is relapse , 0 is censored.}
\item{group}{platelet level: 1 if more than 100 x 10^9 per L, 0 if less}
}
}
\source{
'timereg' package
}
\usage{
BMTplat
}
\description{
The data contain observations of 408 patients treated with HLA-identical sibling
bone marrow transplantation for myelodysplasia. The dataset is essentially a subset of
the 'bmt' data of the 'timereg' package (minor changes were introduced to break the ties).
}
\references{
Li, J., Le-Rademacher, J., & Zhang, M. J. (2014). Weighted comparison of two cumulative incidence functions with R-CIFsmry package. Computer methods and programs in biomedicine, 116(3), 205-214.
}
\keyword{datasets}
