% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BMTtcell}
\alias{BMTtcell}
\title{Bone Marrow Transplant Registry}
\format{
A data frame with 408 rows and 3 variables:
\describe{
\item{time}{time to event since transplant (in months)}
\item{status}{event status, 1 is dead from treatment related causes, 2 is relapse , 0 is censored.}
\item{group}{presence of T-cell depletion: 1 if present, 0 otherwise}
}
}
\source{
'timereg' package
}
\usage{
BMTtcell
}
\description{
The data contain observations of 408 patients treated with HLA-identical sibling
bone marrow transplantation for myelodysplasia. The data is a subset of
the 'bmt' data of the 'timereg' package, which has been slightly modified to break the ties.
}
\references{
Li, J., Le-Rademacher, J., & Zhang, M. J. (2014). Weighted comparison of two cumulative incidence functions with R-CIFsmry package. Computer methods and programs in biomedicine, 116(3), 205-214.
}
\keyword{datasets}
