% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Freireich}
\alias{Freireich}
\title{Acute Leukemia data from Freireich et al (1963)}
\format{
A data frame with 42 rows and 3 variables:
\describe{
\item{time}{time in remission (in weeks)}
\item{status}{event status, 1 is relapse, 0 is censored}
\item{group}{treatment group: 0 (placebo) or 1 (6-MP)}
}
}
\source{
Data listed in Section 5 in Thomas & Grunkemeier (1975) and Section 11 in Gehan (1965)
}
\usage{
Freireich
}
\description{
In this study there were 21 pairs of subjects, and within each pair one subject
received 6-mercaptopurine (6-MP) and one got placebo. The data are right censored.
See also Gehan (1965) and Thomas & Grunkemeier (1975) who used the data as an
illustrative example (ignoring the pairing).
}
\references{
Freireich et al (1963) Blood 21(6):699-716

Gehan (1965) Biometrika 52:203-223

Thomas & Grunkemeier (1975) JASA 70(352): 865-871
}
\keyword{datasets}
