% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melanoma5}
\alias{melanoma5}
\title{Melanoma competing risks data}
\format{
A data frame with 173 rows and 2 variables:
\describe{
\item{time}{time to event (in years)}
\item{status}{event status, 1 is death due to malignant melanoma, 2 is death due to another cause and 0 is censored}
}
}
\source{
'timereg' package
}
\usage{
melanoma5
}
\description{
These competing risks data relate to survival of patients after operation for malignant melanoma collected at Odense University Hospital between 1962 and
1977. The data are a subsample of the 'melanoma' data of the 'timereg' package (patients who had a tumor thickness of less than 5 cm).
}
\references{
Andersen PK, Skovgaard LT (2010) Regression with linear predictors. Springer, Berlin

Drzewiecki K, Andersen PK (1982) Survival with malignant melanoma: a regression analysis of prognostic factors. Cancer 49:2414–2419
}
\keyword{datasets}
