% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.AalenJohansen}
\alias{print.AalenJohansen}
\title{Print function for object of class 'AalenJohansen'}
\usage{
\method{print}{AalenJohansen}(x, digits = 4, method = NULL, ...)
}
\arguments{
\item{x}{an object of class 'AalenJohansen'}

\item{digits}{number of digits to print the results}

\item{method}{either "EL", "Wald" or "both", depending on whether we want to print the results obtained when using empirical likelihood inference (EL), Wald-type inference (Wald) or both. Default is 'NULL', which means that 'method' inherits the value of the corresponding control parameter used when creating the object 'x'.}

\item{...}{Not used}
}
\value{
no return value, called for printing only.
}
\description{
Print function for object of class 'AalenJohansen'
}
\examples{
x <- AalenJohansen(time=melanoma5$time, cause=melanoma5$status, t=4, level=0.95)
print(x, digits=3, method="EL")

}
\author{
Paul Blanche
}
