% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimA100}
\alias{SimA100}
\title{Simulated competing risks data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{time}{time to event}
\item{status}{event status, 1 is main event, 2 is competing event, 0 is censored.}
\item{group}{group (1 or 0)}
}
}
\source{
Simulated data
}
\usage{
SimA100
}
\description{
The data were simulated as described in Blanche & Eriksson (2023), using  scenario A with sample size n=100.
}
\references{
Blanche & Eriksson (2023). Empirical likelihood comparison of absolute risks.
}
\keyword{datasets}
